Changed Example + Moved Directory
